move Brillouin to Model.jl
